% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.chrono_components}
\alias{autoplot.chrono_components}
\title{Plot chronostatistical variance components}
\usage{
\method{autoplot}{chrono_components}(object, ...)
}
\arguments{
\item{object}{A \code{chrono_components} object.}

\item{...}{Reserved for extensions.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot chronostatistical variance components
}
