% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.chrono_data}
\alias{autoplot.chrono_data}
\title{Plot chronological observations}
\usage{
\method{autoplot}{chrono_data}(object, ...)
}
\arguments{
\item{object}{A \code{chrono_data} object.}

\item{...}{Reserved for extensions.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot chronological observations
}
