% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.chrono_limits}
\alias{autoplot.chrono_limits}
\title{Plot progressive engineering limits}
\usage{
\method{autoplot}{chrono_limits}(object, ...)
}
\arguments{
\item{object}{A \code{chrono_limits} object.}

\item{...}{Reserved for extensions.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot progressive engineering limits
}
