% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chRonostatistics-package.R
\docType{package}
\name{chRonostatistics-package}
\alias{chRonostatistics}
\alias{chRonostatistics-package}
\title{chRonostatistics: Chronological Variograms for Process Data}
\description{
Base-R computations for chronological process decomposition.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ClausonGeomet/chRonostatistics-r}
  \item Report bugs at \url{https://github.com/ClausonGeomet/chRonostatistics-r/issues}
}

}
\author{
\strong{Maintainer}: Matt Clauson \email{matthewclauson@outlook.com} (Clauson Geometallurgy Consultancy) [copyright holder]

}
\keyword{internal}
