% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.chrono_components}
\alias{plot.chrono_components}
\title{Base plot method for chronostatistical components}
\usage{
\method{plot}{chrono_components}(x, ...)
}
\arguments{
\item{x}{A \code{chrono_components} object.}

\item{...}{Passed to \code{ggplot2::autoplot()}.}
}
\value{
The plotted \code{ggplot} object, invisibly.
}
\description{
Base plot method for chronostatistical components
}
