% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.chrono_data}
\alias{plot.chrono_data}
\title{Base plot method for chronological observations}
\usage{
\method{plot}{chrono_data}(x, ...)
}
\arguments{
\item{x}{A \code{chrono_data} object.}

\item{...}{Passed to \code{ggplot2::autoplot()}.}
}
\value{
The plotted \code{ggplot} object, invisibly.
}
\description{
Base plot method for chronological observations
}
