% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.chrono_limits}
\alias{plot.chrono_limits}
\title{Base plot method for progressive engineering limits}
\usage{
\method{plot}{chrono_limits}(x, ...)
}
\arguments{
\item{x}{A \code{chrono_limits} object.}

\item{...}{Passed to \code{ggplot2::autoplot()}.}
}
\value{
The plotted \code{ggplot} object, invisibly.
}
\description{
Base plot method for progressive engineering limits
}
