% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.chrono_variogram}
\alias{plot.chrono_variogram}
\title{Base plot method for a chronological semivariogram}
\usage{
\method{plot}{chrono_variogram}(x, ...)
}
\arguments{
\item{x}{A \code{chrono_variogram} object.}

\item{...}{Passed to \code{ggplot2::autoplot()}.}
}
\value{
The plotted \code{ggplot} object, invisibly.
}
\description{
Base plot method for a chronological semivariogram
}
