% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.chrono_components}
\alias{print.chrono_components}
\title{Print chronostatistical components}
\usage{
\method{print}{chrono_components}(x, ...)
}
\arguments{
\item{x}{A \code{chrono_components} object.}

\item{...}{Arguments passed to the data-frame printer.}
}
\value{
\code{x}, invisibly.
}
\description{
Print chronostatistical components
}
