% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.chrono_data}
\alias{print.chrono_data}
\title{Print chronological data}
\usage{
\method{print}{chrono_data}(x, ...)
}
\arguments{
\item{x}{A \code{chrono_data} object.}

\item{...}{Arguments passed to the data-frame printer.}
}
\value{
\code{x}, invisibly.
}
\description{
Print chronological data
}
