% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.chrono_limits}
\alias{print.chrono_limits}
\title{Print progressive engineering limits}
\usage{
\method{print}{chrono_limits}(x, ...)
}
\arguments{
\item{x}{A \code{chrono_limits} object.}

\item{...}{Arguments passed to the data-frame printer.}
}
\value{
\code{x}, invisibly.
}
\description{
Print progressive engineering limits
}
