% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.chrono_variogram}
\alias{print.chrono_variogram}
\title{Print a chronological variogram}
\usage{
\method{print}{chrono_variogram}(x, ...)
}
\arguments{
\item{x}{A \code{chrono_variogram} object.}

\item{...}{Arguments passed to the data-frame printer.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a chronological variogram
}
