% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.chrono_components}
\alias{summary.chrono_components}
\title{Summarise chronostatistical components}
\usage{
\method{summary}{chrono_components}(object, ...)
}
\arguments{
\item{object}{A \code{chrono_components} object.}

\item{...}{Reserved for method compatibility.}
}
\value{
The underlying component data frame.
}
\description{
Summarise chronostatistical components
}
