% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.chrono_data}
\alias{summary.chrono_data}
\title{Summarise chronological data}
\usage{
\method{summary}{chrono_data}(object, ...)
}
\arguments{
\item{object}{A \code{chrono_data} object.}

\item{...}{Reserved for method compatibility.}
}
\value{
A one-row data frame of sample characteristics.
}
\description{
Summarise chronological data
}
