% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.chrono_variogram}
\alias{summary.chrono_variogram}
\title{Summarise a chronological variogram}
\usage{
\method{summary}{chrono_variogram}(object, ...)
}
\arguments{
\item{object}{A \code{chrono_variogram} object.}

\item{...}{Reserved for method compatibility.}
}
\value{
A one-row data frame.
}
\description{
Summarise a chronological variogram
}
